+++ src/RunnableSuiteProperties.ts
@@ -1,18 +1,19 @@
 import * as c2fs from './FSWrapper';
-import { ExecutableConfigFrameworkSpecific } from './ExecutableConfig';
+import { ExecutableConfigFrameworkSpecific, RunTask } from './ExecutableConfig';
 import { TestGrouping } from './TestGroupingInterface';
-import { ResolveRulePair } from './Util';
+import { ResolveRule } from './util/ResolveRule';
 import { TaskPool } from './TaskPool';
 
 export class RunnableSuiteProperties {
   public constructor(
     public readonly name: string | undefined,
     public readonly description: string | undefined,
-    public readonly varToValue: readonly ResolveRulePair[],
+    public readonly varToValue: readonly ResolveRule[],
     public readonly path: string,
     public readonly options: c2fs.SpawnOptions,
     private readonly _frameworkSpecific: ExecutableConfigFrameworkSpecific,
     _parallelizationLimit: number,
+    public readonly runTask: RunTask,
   ) {
     this.parallelizationPool = new TaskPool(_parallelizationLimit);
   }